+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+# Akeeba Panopticon Connector for Joomla 4.x and 5.x version 1.1.0
+
+* ✨ Remote extension installation
+
 # Akeeba Panopticon Connector for Joomla 4.x and 5.x version 1.0.8
+
 * ✨ Support for Joomla 6
 * ✨ PHP 8.5 compatibility